+++ packages/kc_api/crates/kc-api-interface/src/mst_ship_upgrade.rs
@@ -17,7 +17,7 @@ pub(crate) static KCS_MST_SHIP_UPGRADE: Lazy<Mutex<MstShipUpgrades>> = Lazy::new
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
 #[ts(export, export_to = "get_data.ts")]
 pub struct MstShipUpgrades {
-    pub mst_ship_upgrades: HashMap<i64, MstShipUpgrade>,
+    pub mst_ship_upgrades: HashMap<i32, MstShipUpgrade>,
 }
 
 #[derive(
@@ -25,19 +25,19 @@ pub struct MstShipUpgrades {
 )]
 #[ts(export, export_to = "get_data.ts")]
 pub struct MstShipUpgrade {
-    pub api_id: i64,
-    pub api_current_ship_id: i64,
-    pub api_original_ship_id: i64,
-    pub api_upgrade_type: i64,
-    pub api_upgrade_level: i64,
-    pub api_drawing_count: i64,
-    pub api_catapult_count: i64,
-    pub api_report_count: i64,
-    pub api_aviation_mat_count: i64,
-    pub api_arms_mat_count: i64,
-    pub api_tech_count: i64,
-    pub api_sortno: i64,
-    pub api_boiler_count: Option<i64>,
+    pub api_id: i32,
+    pub api_current_ship_id: i32,
+    pub api_original_ship_id: i32,
+    pub api_upgrade_type: i32,
+    pub api_upgrade_level: i32,
+    pub api_drawing_count: i32,
+    pub api_catapult_count: i32,
+    pub api_report_count: i32,
+    pub api_aviation_mat_count: i32,
+    pub api_arms_mat_count: i32,
+    pub api_tech_count: i32,
+    pub api_sortno: i32,
+    pub api_boiler_count: Option<i32>,
 }
 
 impl MstShipUpgrades {
